Updates for AJAX changes in wicket 6.
Make use of TagTester to simplify assertAttribute methods.

Delete some unused files.

+++ trunk/cwm-data/src/main/java/org/cast/cwm/service/ResponseService.java
@@ -365,7 +365,7 @@ public class ResponseService implements IResponseService {
 		dataProvider.getSortState().setPropertySortOrder("sortOrder", SortOrder.ASCENDING);
 		
 		if (index == null)
-			index = dataProvider.size() - 1;
+			index = (int)dataProvider.size() - 1;
 		
 		if (index < 0)
 			throw new IllegalArgumentException("Index cannot be negative.");
